Refactoring common utils out
