+++ test/mina-placeholder-verifier-proof-test.js
@@ -10,76 +10,22 @@ const path = require("path");
 const losslessJSON = require("lossless-json")
 const {BigNumber} = require("ethers");
 const {getNamedAccounts} = hre
+const {loadParamsFromFile, getVerifierParams, getVerifierParamsAccount} = require("./utils/utils")
+
+const stateProofFile = path.resolve(__dirname, "./data/proof_state.bin");
+const baseParamsFile = path.resolve(__dirname, "./data/verifier_params_state_base.json");
+const scalarParamsFile = path.resolve(__dirname, "./data/verifier_params_state_scalar.json");
+const accountParamsFile = path.resolve(__dirname, './data/verifier_params_account.json');
+const accountProofFile = path.resolve(__dirname, "./data/proof_account.bin");
 
 /* global BigInt */
 
 describe('Mina state proof validation tests', function () {
     const {deployments, getNamedAccounts} = hre;
     const {deploy} = deployments;
-
-    function loadParamsFromFile(jsonFile) {
-        const named_params = losslessJSON.parse(fs.readFileSync(jsonFile, 'utf8'));
-        params = {};
-        params.init_params = [];
-        params.init_params.push(BigInt(named_params.modulus.value));
-        params.init_params.push(BigInt(named_params.r.value));
-        params.init_params.push(BigInt(named_params.max_degree.value));
-        params.init_params.push(BigInt(named_params.lambda.value));
-        params.init_params.push(BigInt(named_params.rows_amount.value));
-        params.init_params.push(BigInt(named_params.omega.value));
-        params.init_params.push(BigInt(named_params.D_omegas.length));
-        for (i in named_params.D_omegas) {
-            params.init_params.push(BigInt(named_params.D_omegas[i].value))
-        }
-        params.init_params.push(named_params.step_list.length);
-        for (i in named_params.step_list) {
-            params.init_params.push(BigInt(named_params.step_list[i].value))
-        }
-        params.init_params.push(named_params.arithmetization_params.length);
-        for (i in named_params.arithmetization_params) {
-            params.init_params.push(BigInt(named_params.arithmetization_params[i].value))
-        }
-
-        params.columns_rotations = [];
-        for (i in named_params.columns_rotations) {
-            r = []
-            for (j in named_params.columns_rotations[i]) {
-                r.push(BigInt(named_params.columns_rotations[i][j].value));
-            }
-            params.columns_rotations.push(r);
-        }
-        return params;
-    }
-
-
-    function getVerifierParams() {
-        let params = {}
-        params['proof'] = fs.readFileSync(path.resolve(__dirname, "./data/proof_state.bin"), 'utf8');
-        params['init_params'] = [[24760, 21744], [], []];
-        params['columns_rotations'] = [[], []]
-
-        // For proof 1
-        let base_params = loadParamsFromFile(path.resolve(__dirname, "./data/verifier_params_state_base.json"));
-        params['init_params'][1] = base_params.init_params;
-        params['columns_rotations'][0] = base_params.columns_rotations;
-
-        // For proof 2
-        let scalar_params = loadParamsFromFile(path.resolve(__dirname, "./data/verifier_params_state_scalar.json"));
-        params['init_params'][2] = scalar_params.init_params;
-        params['columns_rotations'][1] = scalar_params.columns_rotations;
-
-        return params;
-    }
-
-    function getVerifierParamsAccount() {
-        let account_path_params = loadParamsFromFile(path.resolve(__dirname, './data/verifier_params_account.json'));
-        account_path_params['proof'] = fs.readFileSync(path.resolve(__dirname, "./data/proof_account.bin"), 'utf8');
-        return account_path_params;
-    }
-
     describe('Ledger Proof - Success', function () {
         it("Should validate correct proof ", async function () {
-            let params = getVerifierParams();
+            let params = getVerifierParams(stateProofFile,baseParamsFile,scalarParamsFile);
             await deployments.fixture(['minaPlaceholderVerifierFixture']);
             let minaPlaceholderVerifier = await ethers.getContract('MinaState');
             let minaPlaceholderVerifierIF = await ethers.getContractAt("IMinaPlaceholderVerifier", minaPlaceholderVerifier.address);
@@ -90,7 +36,7 @@ describe('Mina state proof validation tests', function () {
         });
 
         it("Should update and store proof ", async function () {
-            let params = getVerifierParams();
+            let params = getVerifierParams(stateProofFile,baseParamsFile,scalarParamsFile);
             await deployments.fixture(['minaPlaceholderVerifierFixture']);
             let minaPlaceholderVerifier = await ethers.getContract('MinaState');
             let minaPlaceholderVerifierIF = await ethers.getContractAt("IMinaPlaceholderVerifier", minaPlaceholderVerifier.address);
@@ -101,7 +47,7 @@ describe('Mina state proof validation tests', function () {
         });
 
         it("Should validate previously updated & stored correct proof ", async function () {
-            let params = getVerifierParams();
+            let params = getVerifierParams(stateProofFile,baseParamsFile,scalarParamsFile);
             let minaPlaceholderVerifier = await ethers.getContract('MinaState');
             let minaPlaceholderVerifierIF = await ethers.getContractAt("IMinaPlaceholderVerifier", minaPlaceholderVerifier.address);
             expect(await minaPlaceholderVerifierIF.isValidatedLedgerHash("jwYPLbRQa4X86tSJs1aTzusf3TNdVTj58oyWJQB132sEGUtKHcB", {gasLimit: 30_500_000})).to.equal(true);
@@ -110,7 +56,7 @@ describe('Mina state proof validation tests', function () {
 
     describe('Account Proof - Success', function () {
         it("Should validate correct proof ", async function () {
-            let params_state = getVerifierParams();
+            let params_state = getVerifierParams(stateProofFile,baseParamsFile,scalarParamsFile);
             await deployments.fixture(['minaPlaceholderVerifierFixture']);
             let ledger_hash = 'jwYPLbRQa4X86tSJs1aTzusf3TNdVTj58oyWJQB132sEGUtKHcB';
             let minaPlaceholderVerifier = await ethers.getContract('MinaState');
@@ -121,7 +67,7 @@ describe('Mina state proof validation tests', function () {
                 {gasLimit: 30_500_000})
 
 
-            let params_account = getVerifierParamsAccount();
+            let params_account = getVerifierParamsAccount(accountParamsFile,accountProofFile);
 
             const accountData = {
                 public_key: "B62qre3ersHfzQckNuibViWTGyyKwZseztqrjPjBv6SQF384Rg6ESAy",
@@ -154,7 +100,7 @@ describe('Mina state proof validation tests', function () {
 
     describe("Ledger Proof - Failures", function () {
         it("Should emit event on incorrect proof validation", async function () {
-            let params = getVerifierParams();
+            let params = getVerifierParams(stateProofFile,baseParamsFile,scalarParamsFile);
             await deployments.fixture(['minaPlaceholderVerifierFixture'])
             params['proof'] = '0x4554480000000000000000000000000000000000000000000000000000000000'
             let minaPlaceholderVerifier = await ethers.getContract('MinaState');
@@ -167,7 +113,7 @@ describe('Mina state proof validation tests', function () {
         });
 
         it("Should fail to update and store incorrect proof ", async function () {
-            let params = getVerifierParams();
+            let params = getVerifierParams(stateProofFile,baseParamsFile,scalarParamsFile);
             await deployments.fixture(['minaPlaceholderVerifierFixture']);
             params['proof'] = '0x4554480000000000000000000000000000000000000000000000000000000000'
             let minaPlaceholderVerifier = await ethers.getContract('MinaState');
@@ -179,7 +125,7 @@ describe('Mina state proof validation tests', function () {
         });
 
         it("Should fail for incorrect hash", async function () {
-            let params = getVerifierParams();
+            let params = getVerifierParams(stateProofFile,baseParamsFile,scalarParamsFile);
             await deployments.fixture(['minaPlaceholderVerifierFixture'])
             let minaPlaceholderVerifier = await ethers.getContract('MinaState');
             let minaPlaceholderVerifierIF = await ethers.getContractAt("IMinaPlaceholderVerifier", minaPlaceholderVerifier.address);
@@ -192,7 +138,7 @@ describe('Mina state proof validation tests', function () {
 
     describe('Account Proof - Failures', function () {
         it("Should reject invalid ledger hash", async function () {
-            let params_state = getVerifierParams();
+            let params_state = getVerifierParams(stateProofFile,baseParamsFile,scalarParamsFile);
             await deployments.fixture(['minaPlaceholderVerifierFixture']);
             let ledger_hash = 'jwYPLbRQa4X86tSJs1aTzusf3TNdVTj58oyWJQB132sEGUtKHcB';
             let minaPlaceholderVerifier = await ethers.getContract('MinaState');
@@ -203,7 +149,7 @@ describe('Mina state proof validation tests', function () {
                 {gasLimit: 30_500_000})
 
 
-            let params_account = getVerifierParamsAccount();
+            let params_account = getVerifierParamsAccount(accountParamsFile,accountProofFile);
 
             const accountData = {
                 public_key: "B62qre3ersHfzQckNuibViWTGyyKwZseztqrjPjBv6SQF384Rg6ESAy",
